ST: fix systime type

+++ os/hal/ports/SN32/LLD/SN32F2xx/SysTick/hal_st_lld.h
@@ -144,7 +144,7 @@ static inline systime_t st_lld_get_counter(void) {
  * @notapi
  */
 static inline void st_lld_start_alarm(systime_t abstime) {
-  SN32_ST_TIM->MR0 = (uint32_t)abstime;
+  SN32_ST_TIM->MR0 = (uint16_t)abstime;
   SN32_ST_TIM->IC &= 0x1FFFFFF;
   SN32_ST_TIM->MCTRL = mskCT16_MR0IE_EN;
 }
@@ -166,7 +166,7 @@ static inline void st_lld_stop_alarm(void) {
  * @notapi
  */
 static inline void st_lld_set_alarm(systime_t abstime) {
-  SN32_ST_TIM->MR0 = (uint32_t)abstime;
+  SN32_ST_TIM->MR0 = (uint16_t)abstime;
 }
 
 /**
